Move context utils into new `context_helpers` module.
